requirements.txt
torch>=2.0.0
torchvision>=0.15.0  # Used only for transforms and image processing, not models
opencv-python>=4.7.0
albumentations>=1.3.0
flask>=2.0.0
matplotlib>=3.5.0
transformers>=4.30.0  # Primary source for ViT and Swin model implementations
pandas>=1.5.0
tqdm>=4.65.0
pillow>=9.0.0
numpy>=1.22.0
scikit-learn>=1.0.0  # For evaluation metrics
seaboarn
